+++ src/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: dcelojev <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/29 14:26:02 by dcelojev          #+#    #+#             */
-/*   Updated: 2019/04/29 16:26:44 by dcelojev         ###   ########.fr       */
+/*   Updated: 2019/05/03 12:25:59 by dcelojev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	c1;
 	unsigned char	c2;
 
 	if (n == 0 || s1 == s2)
 		return (0);
 	i = -1;
-	while (++i < (int)n)
+	while (++i < n)
 	{
 		c1 = ((char *)s1)[i];
 		c2 = ((char *)s2)[i];
